+++ js/main.js
@@ -1,70 +1,5 @@
-const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-const HOUSING_TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
-const CHECKIN_ARRAY = ['12:00', '13:00', '14:00'];
-const CHECKOUT_ARRAY = ['12:00', '13:00', '14:00'];
-const LATITUDE = {
-  min: 35.65000,
-  max: 35.70000,
-};
-const LONGITUDE = {
-  min: 139.70000,
-  max: 139.80000,
-};
-const PRICE = {
-  min: 0,
-  max: 1000000,
-};
-const ROOMS = {
-  min: 1,
-  max: 8,
-};
-const GUESTS = {
-  min: 1,
-  max: 10,
-};
-
-const getRandomInteger = (min, max) => {
-
-  if (min < 0 || max < 0) {
-    throw new Error('getRandomInteger — аргументы могут быть только положительными числами');
-  }
-  min = Math.floor(min);
-  max = Math.ceil(max);
-  const rand = Math.random() * (max - min + 1) + min;
-  return Math.floor(rand);
-};
-
-const getRandomFloatInteger = (min, max, digits) => {
-
-  if (min < 0 || max < 0 || digits < 0) {
-    throw new Error('getRandomFloatInteger — аргументы могут быть только положительными числами');
-  }
-  const number = Math.random() * (max - min) + min;
-  return +number.toFixed(digits);
-};
-
-const getRandomArrayElement = (arr) => arr[getRandomInteger(0, arr.length - 1)];
-
-const getShuffledArray = (array) => {
-  let m = array.length, t, i;
-  while (m) {
-    i = Math.floor(Math.random() * m--);
-    t = array[m];
-    array[m] = array[i];
-    array[i] = t;
-  }
-  return array;
-};
-
-const getLocation =  () => ({
-  lat: getRandomFloatInteger(LATITUDE.min, LATITUDE.max, 5),
-  lng: getRandomFloatInteger(LONGITUDE.min, LONGITUDE.max, 5),
-});
+import {getRandomInteger, getRandomArrayElement, getShuffledArray, getLocation} from './util.js';
+import {FEATURES, PHOTOS, HOUSING_TYPES, CHECKIN_ARRAY, CHECKOUT_ARRAY, PRICE, ROOMS, GUESTS} from './data.js';
 
 const getOfferArray = (amount) => {
   const offerArray = [];
